Leart something about scenes
Basically: (1) if you want to load a scene, the first thing you have to do is go to File > Build Settings and drag and drop your scen in the window. In this way, you can call from code 'SceneManager.LoadScene(sceneName)' to go to that scene. (2) if you want a persistent object throughout scenes, you can create a MonoBehaviour (let's call it 'DontDestroy') that calls in the awake function 'DontDestroyOnLoad(gameObject);', so that the object to which this is attached is not destroyed when a new scene is loaded