Cleaned useless code & now it works
I also made the stopAta parameter of the ScriptableObject of the enemy. Anyway, to solve the fact that sometimes it gets stuck when someone moves it during the following of a path, there are a couple of ways to solve it: 1) give up the search if you don't get to the next destination in the amount of time you expect will be necessary to arrive there. 2) re-compute the path if you don't get in time to the next destination. Do it when your brain isn't smoking man